+++ app/Http/Resources/EpisodeResourceBasic.php
@@ -85,7 +85,7 @@ class EpisodeResourceBasic extends JsonResource
 
         // Return the array
         return [
-            'givenRating' => (double) $givenRating?->rating,
+            'givenRating' => (float) $givenRating?->rating,
             'givenReview' => $givenRating?->description,
             'isWatched' => $watchStatus->boolValue
         ];
