+++ app/Http/Resources/GameResourceBasic.php
@@ -86,7 +86,7 @@ class GameResourceBasic extends JsonResource
         // Return the array
         return [
             'library' => [
-                'rating' => (double) $givenRating?->rating,
+                'rating' => (float) $givenRating?->rating,
                 'review' => $givenRating?->description,
                 'status' => $library?->status,
                 'rewatchCount' => $library?->rewatch_count,
